% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engorgio.R
\name{engorgio}
\alias{engorgio}
\title{Engorgio - Spread Key-Value Pairs to Widen Data}
\usage{
engorgio(data, names_from, values_from, ...)
}
\arguments{
\item{data}{A data frame or tibble that you wish to transform into a wider format.}

\item{names_from}{The column in \code{data} that contains the variable names which
will be spread out across the new columns in the wider format.}

\item{values_from}{The column in \code{data} that contains the values to fill the
new columns created by spreading \code{names_from}.}

\item{...}{Additional arguments passed on to \code{tidyr::pivot_wider()}.}
}
\value{
A wider data frame or tibble where key-value pairs specified by
\code{names_from} and \code{values_from} are spread across multiple columns, enlarging
the dataset horizontally.
}
\description{
\code{engorgio} spell expands your data frame, spreading key-value pairs across
multiple columns, thus making the dataset wider. It is akin to the magical
process of enlarging objects, but for data frames, using \code{tidyr::pivot_wider()}.
This function is particularly useful when you want to transform long-format
data into a wide-format, making it easier to analyze and visualize.
}
\examples{
\dontrun{
library(tidyr)
# Assuming a data frame `long_data` with columns: id, category, value
engorgio(long_data, names_from = "category", values_from = "value")
}
}
\seealso{
Other data transformation functions: 
\code{\link{reducio}()},
\code{\link{transfiguro}()}
}
\concept{data transformation functions}
