% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hp_characters.R
\docType{data}
\name{hp_characters}
\alias{hp_characters}
\title{Harry Potter Characters Dataset (hp_characters)}
\format{
A data frame with 15 rows and 5 variables:
\describe{
\item{Name}{character: The name of the character.}
\item{House}{character: The Hogwarts house to which the character belongs (Gryffindor, Slytherin, Ravenclaw, Hufflepuff).}
\item{Role}{character: The character's role or occupation within the wizarding world (e.g., Student, Professor, Headmaster).}
\item{Special_Abilities}{character: Unique magical abilities or skills the character is known for (e.g., Parseltongue, Animagus).}
\item{Patronus}{character: The form of the character's Patronus, if known. 'None' indicates that the Patronus is not known or not applicable.}
}
}
\source{
Derived from the Harry Potter series by J.K. Rowling.
}
\usage{
data(hp_characters)
}
\description{
This dataset contains information on key characters from the Harry Potter series.
It includes their house affiliations, roles within the wizarding world, special abilities, and known Patronuses.
It's designed to provide a comprehensive overview of each character's unique attributes for educational and
analytical purposes, especially for fans of the series looking to explore the data with R.
}
\keyword{datasets}
