% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reducio.R
\name{reducio}
\alias{reducio}
\title{Reducio - Condense Wide Data into Long Format}
\usage{
reducio(data, cols, names_to = "name", values_to = "value", ...)
}
\arguments{
\item{data}{A data frame or tibble that you wish to transform into a longer format.}

\item{cols}{Columns in \code{data} that you want to condense into longer format. Can
be specified using dplyr select helpers like \code{everything()}, \code{starts_with()},
\code{ends_with()}, etc.}

\item{names_to}{The name of the new column in the long format that will contain
the names of the condensed columns. Defaults to \code{"name"}.}

\item{values_to}{The name of the new column in the long format that will contain
the values from the condensed columns. Defaults to \code{"value"}.}

\item{...}{Additional arguments passed on to \code{tidyr::pivot_longer()}.}
}
\value{
A longer, more detailed data frame or tibble where specified wide-format
columns are condensed into two columns: one for the variable names (\code{names_to})
and one for the values (\code{values_to}).
}
\description{
\code{reducio} magically condenses wide-format data into a longer, more detailed
format, mirroring the functionality of \code{tidyr::pivot_longer()}. This function
is ideal for gathering multiple columns into key-value pairs, thereby
stretching the data frame vertically and reducing its width.
}
\examples{
\dontrun{
library(tidyr)
# Assuming a data frame `wide_data` with columns: id,
dermeasurement1, measurement2
reducio(wide_data, cols = starts_with("measurement"),
names_to = "measurement_type",
values_to = "measurement_value")
}
}
\seealso{
Other data transformation functions: 
\code{\link{engorgio}()},
\code{\link{transfiguro}()}
}
\concept{data transformation functions}
