% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transfiguro.R
\name{transfiguro}
\alias{transfiguro}
\title{Transfiguro - Transform or Add New Variables to a Data Frame}
\usage{
transfiguro(.data, ...)
}
\arguments{
\item{.data}{A data frame or tibble to be transformed.}

\item{...}{Additional arguments specifying the transformations to be applied.
Each argument can be a new column name followed by a formula to calculate its values.}
}
\value{
A transformed data frame or tibble with the new or modified variables.
}
\description{
\code{transfiguro} magically transforms or adds new variables to a data frame,
mimicking the functionality of \code{dplyr::mutate()}. Use this function to
seamlessly modify or extend your data frame with new calculations or
transformations, all within the enchanting world of Harry Potter-themed data analysis.
}
\examples{
\dontrun{
library(dplyr)
# Assuming star_data is a data frame with columns: name, house, and year
transfiguro(star_data, age_next_year = year + 1)
}
}
\seealso{
Other data transformation functions: 
\code{\link{engorgio}()},
\code{\link{reducio}()}
}
\concept{data transformation functions}
